@@ -1,85 +1,98 @@
 {
-  "name": "forge",
-  "version": "5.1.2",
+  "name": "jigs",
+  "version": "8.1.0",
   "description": "Pre-configured stack to build static websites.",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mrhubbs/forge.git"
-  },
-  "main": "build/index.js",
+  "main": "launch.js",
   "scripts": {
     "build": "NODE_ENV=production webpack",
     "buildw": "NODE_ENV=development webpack --watch",
     "test": "jest",
     "testw": "jest --watch"
   },
-  "author": "William Sember",
+  "author": "Crown IxD",
   "license": "ISC",
   "babel": {
     "plugins": [
-      "transform-es2015-modules-commonjs"
+      "@babel/plugin-transform-modules-commonjs"
     ]
   },
   "dependencies": {
-    "@babel/core": "^7.4.4",
-    "autoprefixer": "^9.5.1",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.8.0",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-    "browser-sync": "^2.26.5",
-    "chokidar": "^2.1.5",
-    "colorette": "^1.0.7",
-    "copy-webpack-plugin": "^5.0.3",
+    "@babel/core": "^7.7.7",
+    "@babel/plugin-transform-modules-commonjs": "^7.7.5",
+    "autoprefixer": "^9.7.3",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.0.3",
+    "babel-helper-vue-jsx-merge-props": "^2.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-syntax-jsx": "^6.18.0",
+    "babel-plugin-transform-vue-jsx": "^3.7.0",
+    "chokidar": "^2.1.8",
+    "colorette": "^1.1.0",
+    "copy-webpack-plugin": "^5.1.1",
     "crocks": "^0.11.1",
     "css-loader": "^1.0.1",
     "electron-packager": "^13.1.1",
     "eslint": "^5.16.0",
     "eslint-formatter-friendly": "^6.0.0",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-vue": "^5.2.2",
+    "eslint-loader": "^2.2.1",
+    "eslint-plugin-vue": "^5.2.3",
     "file-loader": "^3.0.1",
-    "fs-jetpack": "^2.2.2",
+    "fs-jetpack": "^2.2.3",
     "glob-all": "^3.1.0",
-    "highlight.js": "^9.15.6",
+    "gsap": "^2.1.3",
+    "hash-sum": "^2.0.0",
+    "highlight.js": "^9.17.1",
+    "http-server": "^0.12.0",
     "js-yaml": "^3.13.1",
+    "json-loader": "^0.5.7",
+    "lru-cache": "^5.1.1",
     "markdown-it": "^8.4.2",
-    "markdown-it-anchor": "^5.0.2",
-    "markdown-it-footnote": "^3.0.1",
+    "markdown-it-anchor": "^5.2.5",
+    "markdown-it-footnote": "^3.0.2",
     "markdown-it-ins": "^2.0.0",
     "markdown-it-mark": "^2.0.0",
     "markdown-it-sub": "^1.0.0",
     "markdown-it-sup": "^1.0.0",
-    "markdown-it-table-of-contents": "^0.4.4",
-    "nodemon": "^1.19.0",
-    "postcss": "^7.0.16",
+    "mini-css-extract-plugin": "^0.7.0",
+    "nodemon": "^1.19.4",
+    "postcss": "^7.0.26",
     "postcss-clean": "^1.1.0",
     "postcss-easy-import": "^3.0.0",
     "postcss-extend": "^1.0.5",
     "postcss-loader": "^3.0.0",
-    "postcss-mixins": "^6.2.1",
+    "postcss-mixins": "^6.2.3",
     "postcss-nested": "^3.0.0",
     "postcss-purgecss": "^1.0.0",
     "postcss-simple-vars": "^4.1.0",
-    "purgecss-webpack-plugin": "^1.5.0",
-    "semver": "^6.0.0",
+    "purgecss-webpack-plugin": "^1.6.0",
+    "semver": "^6.3.0",
     "ssh2-sftp-client": "^3.1.0",
     "style-loader": "^0.23.1",
     "tailwindcss": "^0.6.6",
     "url-loader": "^1.1.2",
-    "vue": "^2.6.10",
-    "vue-devtools": "^5.0.0-beta.1",
+    "vue": "^2.6.11",
+    "vue-devtools": "^5.1.3",
     "vue-eslint-parser": "^5.0.0",
-    "vue-loader": "^15.7.0",
+    "vue-jest": "^3.0.5",
+    "vue-loader": "^15.8.3",
     "vue-markdown": "^2.2.4",
-    "vue-server-renderer": "^2.6.10",
-    "vue-template-compiler": "^2.5.22",
-    "webpack": "^4.31.0",
-    "webpack-dev-middleware": "^3.6.2"
+    "vue-meta": "^2.3.1",
+    "vue-router": "^3.1.3",
+    "vue-server-renderer": "^2.6.11",
+    "vue-template-compiler": "^2.6.11",
+    "vuex": "^3.1.2",
+    "vuex-router-sync": "^5.0.0",
+    "webpack": "^4.41.5",
+    "webpack-dev-server": "^3.10.1",
+    "webpack-merge": "^4.2.2",
+    "webpack-virtual-modules": "^0.1.12",
+    "yaml-loader": "^0.5.0"
   },
   "devDependencies": {
-    "jest": "^24.8.0",
-    "webpack-cli": "^3.3.2",
+    "jest": "^24.9.0",
+    "webpack-cli": "^3.3.10",
     "webpack-node-externals": "^1.7.2"
   }
 }
