@@ -1,45 +1,44 @@
 
-const path = require('path')
-const jetpack = require('fs-jetpack')
+import path from 'path'
+import jetpack from 'fs-jetpack'
 
-const Webpack = require('webpack')
-const WebpackMiddleware = require('webpack-dev-middleware')
+import Webpack from 'webpack'
+import webpackMerge from 'webpack-merge'
+// import VirtualModulesPlugin from 'webpack-virtual-modules'
+import WebpackDevServer from 'webpack-dev-server'
 import { curry } from 'crocks'
 import Async from 'crocks/Async'
 
-const unbundledRequire = require('./reusable/unbundledRequire')
-const logging = require('./logging')
+import unbundledRequire from './reusable/unbundledRequire'
+import logging from './logging'
+// import { addEntries } from './builder/jigs-build-webpack/plugin-support'
 
 // Gets the webpack config to use.
-const getWebpackConfig = (mode, forgeConfig) => {
+const getWebpackConfig = (jigsConfig, buildMode, context='client') => {
   // Don't want Webpack to process the default Webpack config.
   // Webpack isn't designed for it's config file to be bundled by Webpack. You
   // can get it to work but I think it's fragile.
-  const makeDefaultWebpackConfig = unbundledRequire('../configs/webpack.config.js').makeConfig
-  const defaultWebpackConfig = makeDefaultWebpackConfig(mode, forgeConfig)
+  const makeDefaultWebpackConfig = unbundledRequire(
+    `../configs/webpack.${context}.config.js`
+  ).makeConfig
+  const defaultWebpackConfig = makeDefaultWebpackConfig(buildMode, jigsConfig)
 
-  const clientWebpackConfigPath = path.join(process.cwd(), 'webpack.config.js')
+  const projectsWebpackConfig = path.join(process.cwd(), `webpack.${context}.config.js`)
 
   // if this project config doesn't exist, then we shouldn't load it
-  if (jetpack.exists(clientWebpackConfigPath)) {
-    const clientWebpackConfig = unbundledRequire(clientWebpackConfigPath)
+  if (jetpack.exists(projectsWebpackConfig)) {
+    const clientWebpackConfig = unbundledRequire(projectsWebpackConfig)
 
     // Overlay the client Webpack config on the default one.
-    // TODO: merge these much more intelligently.
-    return Object.assign({ }, defaultWebpackConfig, clientWebpackConfig)
+    return webpackMerge(defaultWebpackConfig, clientWebpackConfig)
   } else {
     return defaultWebpackConfig
   }
-}
 
-// Determine, by the presence of a webpack config, if the project wants to use
-// webpack.
-export const shouldUse = () => jetpack.exists(
-  path.join(
-    process.cwd(),
-    'webpack.config.js'
-  )
-)
+  // let { virtualModules, pageEntries } = addEntries(jigsConfig)
+  // const virtualModulesInstance = new VirtualModulesPlugin(virtualModules)
+  // config.plugins.push(virtualModulesInstance)
+}
 
 // error handler adapted from webpack example
 const errHandler = curry((reject, resolve, err, stats) => {
@@ -66,7 +65,6 @@ const errHandler = curry((reject, resolve, err, stats) => {
   }
 
   if (success) {
-    logging.logDetail('done')
     resolve()
     return
   }
@@ -74,62 +72,65 @@ const errHandler = curry((reject, resolve, err, stats) => {
   reject('errors or warnings')
 })
 
-// () => Async
-export const build = forgeConfig => {
+const silentErrHandler = curry((reject, resolve, err, stats) => {
+  if (err) {
+    reject(err)
+  } else if (stats.hasErrors) {
+    reject()
+  } else {
+    resolve()
+  }
+})
+
+// (jigsConfig, 'development' | 'production', 'client' | 'server') => Async
+export const build = (jigsConfig, buildMode, context, shouldTalk='you may speak') => {
   return Async((reject, resolve) => {
-    logging.logInfo('building w/ webpack!')
     Webpack(
-      getWebpackConfig('production', forgeConfig)
+      getWebpackConfig(jigsConfig, buildMode, context)
     ).run(
-      errHandler(reject, resolve)
+      (shouldTalk === 'silent') ? silentErrHandler(reject, resolve) : errHandler(reject, resolve)
     )
   })
 }
 
-// only use this in development mode!
-export const getMiddleware = forgeConfig => {
-  const webpackConfig = getWebpackConfig('development', forgeConfig)
-
-  return [
-    WebpackMiddleware(Webpack(webpackConfig), {
-      publicPath: webpackConfig.output ? webpackConfig.output.publicPath : '/',
-      noInfo: false,
-      stats: true,
-      reporter: (middlewareOptions, options) => {
-        const { state, stats } = options
-
-        if (state) {
-          const displayStats = (middlewareOptions.stats !== false)
-
-          if (displayStats) {
-            // get a more displayable form of the stats
-            const info = stats.toJson()
-
-            // show warnings first
-            if (stats.hasWarnings()) {
-              info.warnings.map(w => logging.logFailure(w, { calm: true }))
-            }
-
-            if (stats.hasErrors()) {
-              info.errors.map(e => logging.logFailure(e, { calm: true }))
-            }
-          }
-
-          if (stats.hasErrors()) {
-            logging.logFailure('Failed to compile')
-          } else if (stats.hasWarnings()) {
-            logging.logInfo('Compiled with warnings')
-          } else {
-            logging.logDetail('Compiled successfully')
-          }
-        } else {
-          logging.logDetail('Compiling...')
-        }
+// (jigsConfig, 'development' | 'production', 'client' | 'server') => Async
+export const buildWatch = (jigsConfig, buildMode, context, shouldTalk='you may speak') =>
+  Async((reject, resolve) => {
+    Webpack(
+      getWebpackConfig(jigsConfig, buildMode, context)
+    ).watch(
+      {
+        ignored: [ 'node_modules' ],
+        aggregateTimeout: 3000
       },
-      // Since we're using BrowserSync, we need to "save our work".
-      // TODO: why do we have to do this even though we're incorporated into
-      // browserSync as middleware?
-      writeToDisk: true
-    })
-  ]
-}
+      (shouldTalk === 'silent') ? silentErrHandler(reject, resolve) : errHandler(reject, resolve)
+    )
+  })
+
+// (jigsConfig, 'client' | 'server') => Async
+export const develop = (jigsConfig, buildMode, context) =>
+  Async((reject, resolve) => {
+    const host = jigsConfig.devServer.host
+    const port = jigsConfig.devServer.clientPort
+
+    const config = webpackMerge(
+      getWebpackConfig(jigsConfig, buildMode, context),
+      {
+        watch: true,
+        watchOptions: {
+          ignored: [ 'node_modules' ],
+          aggregateTimeout: 3000
+        }
+      }
+    )
+    const compiler = Webpack(config)
+    // resolve the Async when the build completes
+    compiler.hooks.done.tap('jigs-done-check', resolve)
+
+    const server = new WebpackDevServer(
+      compiler,
+      config.devServer
+    )
+
+    server.listen(port, host)
+  })
