@@ -1,59 +1,74 @@
-// code for HTML building - build mode
 
-import Async from 'crocks/Async'
-import eitherToAsync from 'crocks/Async/eitherToAsync'
-import { identity } from 'crocks/combinators'
+import path from 'path'
 
-import { logInfo, logDetail } from '../logging'
-import { loadLayouts } from './layouts'
-import { loadPages, createPageRenderers, writePages } from './pages'
-import { objectToArray } from '../reusable/functional'
+import jetpack from 'fs-jetpack'
+import { curry, map, compose } from 'crocks'
+
+import {
+  createRenderer,
+  SERVER_BUNDLE_NAME,
+  CLIENT_MANIFEST_NAME
+} from './bundle-renderer'
+import { logInfo, formatDetail, relogLine } from '../logging'
 import { deletePathIfSafe } from '../reusable/toolbox'
 
-// ForgeConfig -> { Layout}(optional) -> [ Page ](optional) -> Async(Error String, ())
-// run a production build
-// builds HTML and copies assets
-export const build = (config, layouts, pages) => {
-  let createRenderers = Async.of(createPageRenderers(config))
-  .ap(eitherToAsync(layouts || loadLayouts(config)))
-  .ap(eitherToAsync(pages || loadPages(config)))
-  // We've wrapped `createPageRenderers` in an Async so we can conveniently load
-  // the layouts and pages concurrently and apply them to it. However, that
-  // function ultimately returns an Async. So we've got nested Asyncs. To
-  // simplify running the Async we'll smish them together.
-  // NOTE: normally we'd use a method called `join`. For some reason Crocks
-  // doesn't have that. However, it does have chain, which, by definition,
-  // performs an operation and than flattens the Monad.
-  .chain(identity)
+// JigsConfig -> Async(Error String, String)
+export const cleanBuild = jigsConfig => {
+  logInfo('cleaning build')
+  return deletePathIfSafe(jigsConfig, jigsConfig.dirs.buildRoot)
+}
 
-  // TODO: unfortutunately, we get a nested Async structure
-  // create the renderers
-  // TODO: move into pages.js?
-  return Async((reject, resolve) => {
-    createRenderers.fork(
-      reject,  // woops, error creating the renderers or rendering
-      // creating renderers worked and we rendered
-      // we've rendered and gotten all the pages, now write them all
-      pages => {
-        // write pages
-        writePages(config, objectToArray(pages))
-        .fork(
-          reject,  // woops, error writing pages
-          // all done!
-          pageResults => {
-            logInfo('Rendered pages')
-            pageResults.forEach(logDetail)
+const loadServerFile = jigsConfig => compose(
+  JSON.parse,
+  jetpack.read,
+  x => path.resolve(jigsConfig.dirs.build, x)
+)
 
-            resolve()
+// use the bundle renderer to generate and write pages
+export const generateAndWritePages = curry((jigsConfig, routes) => {
+  // create a bundle renderer
+  const renderer = createRenderer(
+    jigsConfig,
+    loadServerFile(jigsConfig)(SERVER_BUNDLE_NAME),
+    loadServerFile(jigsConfig)(CLIENT_MANIFEST_NAME)
+  )
+
+  // for each page...
+  map(
+    ([ routePath, routeFile]) => {
+      const context = { url: routePath }
+
+      // render to html
+      renderer.renderToString(context, (err, html) => {
+        if (err) {
+          console.error(err.message)
+        } else {
+          // process the filepath
+          // change a filepath like foo/bar.html to foo/bar/index.html
+          const spl = routeFile.split(path.sep)
+          let filePath
+          if (spl.length === 1) {
+            filePath = spl[spl.length - 1].split('.')[0] + '.html'
+          } else {
+            filePath = path.join(
+              spl.slice(0, -1).join(path.sep),
+              spl[spl.length - 1].split('.')[0],
+              'index.html'
+            )
           }
-        )
-      }
-    )
-  })
-}
 
-// ForgeConfig -> Async(Error String, String)
-export const cleanBuild = config => {
-  logInfo('Cleaning build')
-  return deletePathIfSafe(config, config.dirs.build)
-}
+          const absFilePath = path.resolve(
+            jigsConfig.dirs.build,
+            filePath
+          )
+          relogLine(formatDetail(filePath))
+          // create the directory the .html file will reside in, if necessary
+          if (spl.length > 1) jetpack.dir(path.dirname(absFilePath))
+          // write the file
+          jetpack.write(absFilePath, html)
+        }
+      })
+    },
+    routes
+  )
+})
