@@ -17,15 +17,15 @@ const linter = new CLIEngine(
   )
 )
 
-import jetpack from 'fs-jetpack'
 import glob from 'glob-all'
 import yaml from 'js-yaml'
-import { curry, compose, map, chain, bimap, identity } from 'crocks'
+import { curry } from 'crocks'
 import { Left, Right } from 'crocks/Either'
 
-import { LAYOUT_EXT } from './layouts'
+import { writeFrontMatter, writeBlock } from '../reusable/docBlocks'
 
-const LAYOUT_ENTRY_POINT = '<!-- forge-page-contents -->'
+export const LAYOUT_EXT = 'vue'
+const LAYOUT_ENTRY_POINT = '<!-- jigs-page-contents -->'
 
 // String -> [ String ] -> [ String ]
 export const getPaths = (rootPath, exts) => {
@@ -34,12 +34,18 @@ export const getPaths = (rootPath, exts) => {
   return glob.sync(paths)
 }
 
-// String -> String -> Either(Error String, { String, String })
+// { } -> { }
+export const processFrontMatter = frontMatter => ({
+  ...frontMatter,
+  layout: frontMatter.layout ? `${frontMatter.layout}.${LAYOUT_EXT}` : frontMatter.layout
+})
+
+// Blocks -> String -> Either(Error String, Blocks)
 // Removes front matter, if any, from the template and splits it out.
 //
 // NOTE require at the least the front matter intro and outro (---, ---).
 // This helps with detecting errors w/ mistyped front matter intro and outro
-export const extractFrontMatter = curry(rawTemplate => {
+export const extractFrontMatter = curry((blocks, rawTemplate) => {
   // TODO: is this a very slow way to extract the non-front-matter part of the
   // template?
   // const match = rawTemplate.match(/^---([\S\s]*)+?---([\S\s]*)/m)
@@ -47,38 +53,38 @@ export const extractFrontMatter = curry(rawTemplate => {
 
   // TODO: unit-test and optimize algorithm, move out of this function
   if (rawTemplate.startsWith('---\n')) {
-    let template
-    let match = null
-    const lines = rawTemplate.split('\n')
+    let fmLines = null
+    let templateLines = null
+    const sourceLines = rawTemplate.split('\n')
 
     // read through the rawTemplate, looking for the --- at the start of a line
-    for (let i = 1; i < lines.length; i = i + 1) {
-      if (lines[i] === '---') {
+    for (let i = 1; i < sourceLines.length; i = i + 1) {
+      if (sourceLines[i] === '---') {
         // we found the end of the front matter
-        match = lines.slice(1, i).join('\n')
-        template =
-          '\n'.repeat(i + 1) +  // add blank lines so our linter's line no. reports are accurate to the file
-          lines.slice(i + 1).join('\n')
-
+        fmLines = sourceLines.slice(0, i + 1)
+        templateLines = sourceLines.slice(i + 1)
         break
       }
     }
 
-    if (match === null) {
+    if (fmLines === null) {
       return Left(new Error('Cannot find front matter'))
     }
 
     let frontMatter
     try {
-      frontMatter = yaml.safeLoad(match)  // could be undefined
+      frontMatter = yaml.safeLoad(fmLines.slice(1, -1).join('\n'))
+      // could be undefined if the input was an empty string, make sure it's at
+      // least an empty object
+      frontMatter = frontMatter ? frontMatter : { }
     } catch (exc) {
       return Left(exc)
     }
 
-    return Right({
-      frontMatter: frontMatter ? frontMatter : { },  // if undefined, default to empty Object
-      template
-    })
+    let resBlocks = writeFrontMatter(blocks, frontMatter, fmLines)
+    resBlocks = writeBlock(resBlocks, 'template', templateLines)
+
+    return Right(resBlocks)
   } else {
     // no front matter... oh man...
     // TODO: return an Error
@@ -86,29 +92,9 @@ export const extractFrontMatter = curry(rawTemplate => {
   }
 })
 
-// String -> String -> Either(Error String, Layout | Page)
-// loads page or layout
-export const loadResource = curry((rootPath, resPath) =>
-  compose(
-    bimap(
-      err => `${err}\nfile: "${path.relative(rootPath, resPath)}"`,
-      identity
-    ),
-    map(
-      ({ template, frontMatter }) => ({
-        name: path.relative(rootPath, resPath),
-        frontMatter,
-        template
-      })
-    ),
-    chain(lintTemplate(resPath)),
-    map(processFrontMatter),
-    extractFrontMatter
-  )(jetpack.read(resPath))
-)
-
 // String -> String -> null | Error String
 // lint a template
+// eslint-disable-next-line no-unused-vars
 const lintTemplate = curry((filePath, { template, frontMatter }) => {
   let toLint = template
 
@@ -126,19 +112,7 @@ const lintTemplate = curry((filePath, { template, frontMatter }) => {
   return Right({ template, frontMatter })
 })
 
-// { String, String } -> { String, String }
-const processFrontMatter = ({ template, frontMatter }) => {
-  // Add extension to layout names specified in front matter.
-  // It keeps the layouts cleaner if they don't need to have extensions.
-  if (frontMatter.layout) {
-    frontMatter.layout = `${frontMatter.layout}.${LAYOUT_EXT}`
-  }
-
-  return { template, frontMatter }
-}
-
-
-// Layout || Page -> Layout -> Layout
+// Layout | Page -> Layout -> Layout
 // Collapses the given layout into the given parent, merging the front matter
 // and wrapping the parent layout around the child one.
 // Returns a new layout.
@@ -165,11 +139,3 @@ export const transformExtension = curry((is, to, filePath) =>
     filePath.slice(0, -is.length) + to :
     filePath
 )
-
-export const sourcePagePathToBuildPagePath = curry((config, filePath) => {
-  const buildPath = path.join(
-    config.dirs.build, path.relative(config.dirs.pages, filePath)
-  )
-
-  return transformExtension('md', 'html', buildPath)
-})
