@@ -1,9 +1,11 @@
 
-import { build, cleanBuild } from './build'
-import { prototype } from './prototype'
+export {
+  writeGeneratedRoutes,
+  generateRoutes,
+  filePathToRoutePath
+} from './routes'
 
 export {
-  build,
   cleanBuild,
-  prototype
-}
+  generateAndWritePages
+} from './build'
