@@ -11,11 +11,11 @@ const packagerConfigPath = path.join(process.cwd(), 'e-packager.config.js')
 const packagerConfig = jetpack.exists(packagerConfigPath) ? unbundledRequire(packagerConfigPath) : null
 
 if (!packagerConfig) {
-  module.exports = (/* forgeConfig */) => {
+  module.exports = (/* jigsConfig */) => {
     logging.logFailure('Cannot run the Electron packager without an `e-packager.config.js` file')
   }
 } else {
-  module.exports = forgeConfig => {
+  module.exports = jigsConfig => {
     // create a minimal version of the package.json file
     const p = unbundledRequire(path.join(process.cwd(), 'package.json'))
 
@@ -31,7 +31,7 @@ if (!packagerConfig) {
     // write the file out...
     // TODO: use jetpack?
     fs.writeFileSync(
-      path.join(forgeConfig.dirs.build, 'package.json'),
+      path.join(jigsConfig.dirs.build, 'package.json'),
       JSON.stringify(outPgk, null, 2)
     )
 
