@@ -1,52 +1,73 @@
 
-import { curry } from 'crocks'
+import { compose } from 'crocks'
 import { red, green, cyan, bold, gray } from 'colorette'
 
 // TODO: still figuring out the different logging options that should be
 // available and the visual hierarchy
 
-export const logSuccess = msg => {
-  const prefix = bold(green('✓ '))
+const CLEAR_LINE = '\u001b[0K'
+const MOVE_LEFT = '\u001b[1000D'
+const MOVE_UP = '\u001b[1A'
+const RESET_LINE = MOVE_UP + MOVE_LEFT + CLEAR_LINE
 
-  console.log(prefix + green(msg))
+export const writeWithNewline = x => {
+  process.stdout.write('\n' + x)
+}
+export const writeWithoutNewline = x => {
+  process.stdout.write('\n' + x)
 }
 
-// opts -> {
-//   calm: true | false
-// }
-export function logFailure(msg, opts) {
-  const prefix = bold(red('✘ '))
+export const formatHeader = msg =>
+  bold(cyan(`|| ${msg}`))
 
+const INFO_PREFIX = bold('>')
+export const formatInfo = msg =>
+  cyan(`${INFO_PREFIX} ${msg}`)
+
+export const formatDetail = msg =>
+  `  ✓ ${bold(gray(msg))}`
+
+const SUCCESS_PREFIX = bold('✓')
+export const formatSuccess = msg =>
+  green(`${SUCCESS_PREFIX} ${msg}`)
+
+const FAILURE_PREFIX = bold(red('✘'))
+export function formatFailure(msg, opts) {
   if (!opts) opts = { calm: true }
   if (opts.calm === undefined) opts.calm = true
 
   if (!opts.calm) msg = red(msg)
-  console.log(prefix + msg)
+  return `${FAILURE_PREFIX} ${msg}`
 }
 
-export function logInfo(msg) {
-  const prefix = bold(cyan('└ '))
+export const logInfo = compose(writeWithNewline, formatInfo)
+export const logDetail = compose(writeWithNewline, formatDetail)
+export const logHeader = compose(writeWithNewline, formatHeader)
+export const logSuccess = compose(writeWithNewline, formatSuccess)
+export const logFailure = compose(writeWithNewline, formatFailure)
 
-  console.log(prefix + cyan(msg))
+export function relogLine(msg) {
+  writeWithoutNewline(RESET_LINE + msg)
 }
-
-export function logDetail(msg) {
-  console.log(`  ✓ ${bold(gray(msg))}`)
-}
-
-export function logRawDetail(prefix, msg) {
-  console.log(`  ${prefix} ${bold(gray(msg))}`)
-}
-
-export function logHeader(msg) {
-  console.log(bold(cyan(`# ${bold(msg)}`)))
-}
-
-export function logRaw(msg) {
-  console.log(gray(`>> ${msg}`))
-}
-
-export const indentMsg = curry((indent, msg) => indent + msg)
 
 // creates a function to log with the given message, when called
 export const makeLogCall = (logFunc, msg) => () => logFunc(msg)
+
+export default {
+  logHeader,
+  logInfo,
+  logDetail,
+  logSuccess,
+  logFailure,
+
+  formatHeader,
+  formatInfo,
+  formatDetail,
+  formatSuccess,
+  formatFailure,
+
+  makeLogCall,
+  relogLine,
+  writeWithNewline,
+  writeWithoutNewline
+}
